// src/api/model/sinMoradorDto.ts
/**
 * Service Auth
 * API dashboard
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface SinMoradorDto { 
    empresa: string;
    nombreTecnico: string;
    fecha: string;
    geoTecnico: string;
    geoOt: string;
    imagen: string;
    rutTecnico: string;
    numeroOrden: string;
}