// src/api/model/requestSinMoradorDto.ts
/**
 * Service Auth
 * API dashboard
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface RequestSinMoradorDto { 
    limit: number;
    page: number;
    from: number;
    to: number;
    user: string;
    eps: string;
}